Refactoring menu style
